+++ swing/common-ui/src/main/java/is/codion/swing/common/ui/component/table/ColumnSelectionPanel.java
@@ -34,7 +34,6 @@ import java.awt.event.ActionEvent;
 import java.awt.event.ActionListener;
 import java.awt.event.FocusAdapter;
 import java.awt.event.FocusEvent;
-import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.LinkedHashMap;
 import java.util.List;
@@ -88,7 +87,7 @@ final class ColumnSelectionPanel<C> extends JPanel {
 				columnModel.visible(tableColumn.identifier()).set(false);
 			}
 		});
-		new ArrayList<>(columnModel.hidden().columns()).forEach(tableColumn -> {
+		columnModel.hidden().columns().forEach(tableColumn -> {
 			if (visibleStates.get(tableColumn).get()) {
 				columnModel.visible(tableColumn.identifier()).set(true);
 			}