+++ changelog.md
@@ -30,6 +30,7 @@ Codion Change Log
 - ConditionPanel.ConditionState renamed ConditionView.
 - FilterColumnConditionPanel.Fields added.
 - FilterTableColumnModel.Visible and Hidden added, related refactoring.
+- FilterTableColumnModel.containsColumn() and resetColumns() renamed contains() and reset() respectively.
 ### is.codion.framework.domain
 - Entity.copy(), copyBuilder() and deepCopy() moved to Entity.Copy, renamed mutable(), builder() and deep() respectively, immutable() added.
 - ImmutableEntity bug fixed, StackOverflowError when creating an immutable entity with cyclical foreign key dependencies, some refactoring.